# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zen-AI-Pentest API
  description: |
    Enterprise-grade Pentesting Framework API
    
    ## Features
    - **Scan Management**: Create, monitor, and manage security scans
    - **Findings Tracking**: Track and manage security findings
    - **SIEM Integration**: Export to Splunk, Elastic, QRadar, Rapid7
    - **Compliance**: Map findings to ISO 27001, NIST, SOC 2, GDPR
    - **Reporting**: Generate reports in multiple formats
    
    ## Authentication
    API uses Bearer token authentication. Include token in header:
    ```
    Authorization: Bearer YOUR_TOKEN
    ```
  version: 1.0.0
  contact:
    name: Zen-AI-Pentest Team
    url: https://github.com/SHAdd0WTAka/Zen-Ai-Pentest
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zen-ai-pentest.example.com/v1
    description: Production
paths:
...
  /scans:
    get:
      summary: List scans
      description: List all scans with optional filtering
      tags:
        - Scans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: scan_type
          in: query
          schema:
            type: string
            enum: [network, web, cloud, api, container, full]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScanResponse'
    
    post:
      summary: Create scan
      description: Create and start a new security scan
      tags:
        - Scans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanCreate'
      responses:
        '201':
          description: Scan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
  /scans/{scan_id}:
    get:
      summary: Get scan
      description: Get scan details by ID
      tags:
        - Scans
      parameters:
        - name: scan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanDetail'
        '404':
          description: Scan not found
...
  /compliance/scan:
    post:
      summary: Compliance scan
      description: Run compliance scan against target
      tags:
        - Compliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceScanRequest'
      responses:
        '200':
          description: Scan results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceResponse'
...
components:
  schemas:
...
    ScanCreate:
      type: object
      required:
        - target
      properties:
        target:
          type: string
          example: https://example.com
        scan_type:
          type: string
          enum: [network, web, cloud, api, container, full]
          default: network
        options:
          type: object
        tags:
          type: array
          items:
            type: string
    ScanResponse:
      type: object
      properties:
        id:
          type: string
        target:
          type: string
        scan_type:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        findings_count:
          type: integer
    ScanDetail:
      allOf:
        - $ref: '#/components/schemas/ScanResponse'
        - type: object
          properties:
            options:
              type: object
            results:
              type: object
...
    ComplianceScanRequest:
      type: object
      required:
        - target
        - frameworks
      properties:
        target:
          type: string
        frameworks:
          type: array
          items:
            type: string
            enum: [nist_800_53, nist_csf, iso_27001, soc_2, gdpr, pci_dss, hipaa]
        options:
          type: object
    ComplianceResponse:
      type: object
      properties:
        scan_id:
          type: string
        frameworks:
          type: array
          items:
            type: string
        overall_score:
          type: number
        findings_count:
          type: integer
        controls:
          type: array
          items:
            type: object
...