# api/v1/endpoints/compliance.py
#!/usr/bin/env python3
"""
Compliance Endpoints
"""

from fastapi import APIRouter, HTTPException, Query
from pydantic import BaseModel
from typing import List, Dict, Any, Optional
from enum import Enum

router = APIRouter()


class FrameworkType(str, Enum):
    nist_800_53 = "nist_800_53"
    nist_csf = "nist_csf"
    iso_27001 = "iso_27001"
    soc_2 = "soc_2"
    gdpr = "gdpr"
    pci_dss = "pci_dss"
    hipaa = "hipaa"


class ComplianceScanRequest(BaseModel):
    target: str
    frameworks: List[FrameworkType]
    options: Optional[Dict[str, Any]] = None


class ComplianceReportRequest(BaseModel):
    findings: List[Dict[str, Any]]
    frameworks: List[FrameworkType]
    formats: List[str] = ["json", "html"]


class ComplianceResponse(BaseModel):
    scan_id: str
    frameworks: List[str]
    overall_score: float
    findings_count: int
    controls: List[Dict[str, Any]]


@router.post("/scan", response_model=ComplianceResponse)
async def run_compliance_scan(request: ComplianceScanRequest):
    """
    Run compliance scan against target
    
    Maps security findings to compliance frameworks:
    - **NIST 800-53**: US federal security controls
    - **ISO 27001**: Information security management
    - **SOC 2**: Service organization controls
    - **GDPR**: EU data protection
    - **PCI DSS**: Payment card security
    """
    # Mock response - real implementation would scan
    return ComplianceResponse(
        scan_id="scan-12345",
        frameworks=[f.value for f in request.frameworks],
        overall_score=73.5,
        findings_count=12,
        controls=[
            {
                "id": "A.5.17",
                "name": "Authentication Information",
                "framework": "ISO 27001",
                "status": "partial",
                "score": 50
            }
        ]
    )


@router.post("/report")
async def generate_compliance_report(request: ComplianceReportRequest):
    """
    Generate compliance report
    
    Export formats: json, html, md, csv
    """
    return {
        "reports": [
            {
                "format": fmt,
                "url": f"/reports/compliance-{fmt}",
                "size": "245 KB"
            }
            for fmt in request.formats
        ],
        "frameworks": [f.value for f in request.frameworks],
        "generated_at": "2026-02-04T06:00:00Z"
    }


@router.get("/frameworks")
async def list_frameworks():
    """List supported compliance frameworks"""
    return {
        "frameworks": [
            {
                "id": "nist_800_53",
                "name": "NIST SP 800-53 Rev 5",
                "description": "Security and Privacy Controls for Federal Information Systems",
                "controls_count": 1086
            },
            {
                "id": "iso_27001",
                "name": "ISO/IEC 27001:2022",
                "description": "Information Security Management Systems",
                "controls_count": 93
            },
            {
                "id": "soc_2",
                "name": "SOC 2 Type II",
                "description": "Service Organization Control 2",
                "controls_count": 64
            },
            {
                "id": "gdpr",
                "name": "GDPR",
                "description": "EU General Data Protection Regulation",
                "controls_count": 99
            },
            {
                "id": "pci_dss",
                "name": "PCI DSS 4.0",
                "description": "Payment Card Industry Data Security Standard",
                "controls_count": 78
            }
        ]
    }


@router.get("/mappings/{finding_type}")
async def get_control_mappings(
    finding_type: str,
    frameworks: List[FrameworkType] = Query(default=[FrameworkType.iso_27001])
):
    """Get compliance control mappings for finding type"""
    return {
        "finding_type": finding_type,
        "mappings": [
            {
                "framework": f.value,
                "controls": [
                    {"id": "A.5.17", "name": "Authentication"},
                    {"id": "A.8.24", "name": "Cryptography"}
                ]
            }
            for f in frameworks
        ]
    }