# api/v1/endpoints/scans.py
#!/usr/bin/env python3
"""
Scan Management Endpoints
"""

from fastapi import APIRouter, HTTPException, BackgroundTasks, Query
from pydantic import BaseModel, Field
from typing import List, Dict, Any, Optional
from enum import Enum
from datetime import datetime

router = APIRouter()


class ScanType(str, Enum):
    network = "network"
    web = "web"
    cloud = "cloud"
    api = "api"
    container = "container"
    full = "full"


class ScanStatus(str, Enum):
    pending = "pending"
    running = "running"
    completed = "completed"
    failed = "failed"
    cancelled = "cancelled"


class ScanCreate(BaseModel):
    target: str = Field(..., description="Target URL/IP to scan")
    scan_type: ScanType = Field(default=ScanType.network)
    options: Optional[Dict[str, Any]] = Field(default={}, description="Scan options")
    tags: Optional[List[str]] = []


class ScanResponse(BaseModel):
    id: str
    target: str
    scan_type: str
    status: ScanStatus
    progress: int = Field(0, ge=0, le=100)
    created_at: datetime
    started_at: Optional[datetime] = None
    completed_at: Optional[datetime] = None
    findings_count: int = 0
    tags: List[str] = []


class ScanDetail(ScanResponse):
    options: Dict[str, Any]
    results: Optional[Dict[str, Any]] = None
    error_message: Optional[str] = None


# In-memory storage (replace with database in production)
scans_db: Dict[str, Dict[str, Any]] = {}


@router.post("", response_model=ScanResponse, status_code=201)
async def create_scan(scan: ScanCreate, background_tasks: BackgroundTasks):
    """
    Create and start a new security scan
    
    Scan types:
    - **network**: Port scanning, service detection
    - **web**: Web vulnerability scanning (OWASP Top 10)
    - **cloud**: AWS/Azure/GCP security scanning
    - **api**: API security testing
    - **container**: Docker/Kubernetes security
    - **full**: Comprehensive all-in-one scan
    """
    scan_id = f"scan-{datetime.now().strftime('%Y%m%d-%H%M%S')}-{len(scans_db)}"
    
    scan_data = {
        "id": scan_id,
        "target": scan.target,
        "scan_type": scan.scan_type.value,
        "status": "pending",
        "progress": 0,
        "created_at": datetime.now(),
        "options": scan.options,
        "tags": scan.tags,
        "findings_count": 0
    }
    
    scans_db[scan_id] = scan_data
    
    # In production: trigger actual scan in background
    # background_tasks.add_task(run_scan, scan_id, scan)
    
    return ScanResponse(**scan_data)


@router.get("", response_model=List[ScanResponse])
async def list_scans(
    status: Optional[ScanStatus] = None,
    scan_type: Optional[ScanType] = None,
    limit: int = Query(20, ge=1, le=100),
    offset: int = Query(0, ge=0)
):
    """List all scans with optional filtering"""
    results = list(scans_db.values())
    
    if status:
        results = [s for s in results if s["status"] == status.value]
    if scan_type:
        results = [s for s in results if s["scan_type"] == scan_type.value]
    
    # Sort by created_at desc
    results.sort(key=lambda x: x["created_at"], reverse=True)
    
    return [ScanResponse(**s) for s in results[offset:offset + limit]]


@router.get("/{scan_id}", response_model=ScanDetail)
async def get_scan(scan_id: str):
    """Get scan details by ID"""
    if scan_id not in scans_db:
        raise HTTPException(status_code=404, detail="Scan not found")
    
    return ScanDetail(**scans_db[scan_id])


@router.post("/{scan_id}/cancel")
async def cancel_scan(scan_id: str):
    """Cancel a running scan"""
    if scan_id not in scans_db:
        raise HTTPException(status_code=404, detail="Scan not found")
    
    scan = scans_db[scan_id]
    if scan["status"] not in ["pending", "running"]:
        raise HTTPException(status_code=400, detail=f"Cannot cancel scan in {scan['status']} state")
    
    scan["status"] = "cancelled"
    return {"message": "Scan cancelled", "scan_id": scan_id}


@router.delete("/{scan_id}")
async def delete_scan(scan_id: str):
    """Delete a scan"""
    if scan_id not in scans_db:
        raise HTTPException(status_code=404, detail="Scan not found")
    
    del scans_db[scan_id]
    return {"message": "Scan deleted", "scan_id": scan_id}


@router.get("/{scan_id}/results")
async def get_scan_results(scan_id: str):
    """Get detailed scan results"""
    if scan_id not in scans_db:
        raise HTTPException(status_code=404, detail="Scan not found")
    
    return {
        "scan_id": scan_id,
        "results": {
            "hosts_discovered": 5,
            "services_found": 12,
            "vulnerabilities": {
                "critical": 1,
                "high": 3,
                "medium": 5,
                "low": 8
            }
        }
    }